requirements.txt
git+https://github.com/Namokarsingh09/Telethon
python-decouple

